[refactor] UI 비즈니스 구분.
